- settings menu update
